% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{pal_functions}
\alias{pal_functions}
\title{Output a character vector containing code for a \code{ggplot2} scale}
\usage{
pal_functions(
  palette,
  which = NULL,
  type = c("discrete", "continuous"),
  reverse = FALSE
)
}
\arguments{
\item{palette}{a \verb{[wacolors]} palette or palette name.}

\item{which}{if not \code{NULL}, the indices or names of a subset of colors to use.}

\item{reverse}{\code{TRUE} if the colors should be reversed.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}},
\code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}, or \code{\link[ggplot2:binned_scale]{ggplot2::binned_scale()}} to control
name, limits, breaks, labels and so forth.}
}
\description{
Call this function to get the code for the \verb{scale_*} functions for a palette.
If using RStudio, the code will be loaded at the console prompt;
otherwise, it will be printed at the terminal. Assumes that \code{ggplot2} has
been loaded into the namespace, or will be by the time the scales are used.
}
\examples{
pal_functions("rainier")

}
