% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{palette_html}
\alias{palette_html}
\title{Create HTML code to view palette}
\usage{
palette_html(x, maxwidth = "30em", height = "1em")
}
\arguments{
\item{x}{the \code{palette} object}

\item{maxwidth}{the maximum width of the output, in CSS units}

\item{height}{the height width of the output, in CSS units}
}
\description{
Create HTML code to view palette
}
