% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{wa_pal}
\alias{wa_pal}
\title{Washington State Color Palette Generator}
\usage{
wa_pal(name, n, type = c("discrete", "continuous"), rev = FALSE)
}
\arguments{
\item{name}{The name of the palette (partial matching supported), or an
actual palette from \verb{[wacolors]}.}

\item{n}{The number of colors in the palette. If this exceeds the actual
number and \code{type} is not provided, it will be set to \code{continuous}.}

\item{type}{Either \code{continuous} or \code{discrete}. Use \code{continuous} if you want
to automatically interpolate between colors.}

\item{rev}{\code{TRUE} if palette should be reversed.}
}
\value{
A vector of colors of type \code{palette}
}
\description{
Generate \code{palette} objects from the \code{wacolors} list
}
\examples{
wa_pal("rainier")
wa_pal(wacolors$rainier)

}
