% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{wa_palette}
\alias{wa_palette}
\title{Washington State Color Palette Generator}
\usage{
wa_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{The name of the palette. Quasiquoted and partial matching
supported.}

\item{n}{The number of colors in the palette. If this exceeds the actual
number and \code{type} is not provided, it will be set to \code{continuous}.}

\item{type}{Either \code{continuous} or \code{discrete}. Use \code{continuous} if you want
to automatically interpolate between colors.}
}
\value{
A vector of colors of type \code{palette}
}
\description{
Generate \code{palette} objects from the \code{wacolors} list
}
