% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wacolors}
\alias{wacolors}
\title{Washington State Color Palettes}
\format{
A list of character vectors containing the color palettes. Some
vectors may contain names for each color.
}
\usage{
wacolors
}
\description{
A collection of colorblind-friendly color palettes for various settings in
the state of Washington.
}
\details{
Available palettes (use \code{\link[=wa_pal]{wa_pal()}} to visualize):
\itemize{
\item \code{rainier}
\item \code{rainier_meadow}
\item \code{larch}
\item \code{forest}
\item \code{understory}
\item \code{seattle_night}
\item \code{seattle_fall}
\item \code{pike_place}
\item \code{fort_worden}
\item \code{chambers_bay}
\item \code{stuart}
\item \code{sound_sunset}
\item \code{washington_pass}
\item \code{winthrop}
\item \code{ross_lake}
\item \code{winter_mountain}
\item \code{library}
\item \code{ferries}
\item \code{volcano}
\item \code{locks}
\item \code{deception_pass}
\item \code{scrubland}
\item \code{palouse}
\item \code{grand_coulee}
\item \code{chihuly}
\item \code{plane_view}
\item \code{UW}
\item \code{flag}
\item \code{tartan}
}
}
\examples{
wacolors$rainier
wacolors$seattle_night[1:2]

}
\keyword{datasets}
