% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{match_pal}
\alias{match_pal}
\title{Takes name of palette or palette itself and returns palette + name}
\usage{
match_pal(name)
}
\description{
Takes name of palette or palette itself and returns palette + name
}
