% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wacolors}
\alias{wacolors}
\title{Washington State Color Palettes}
\format{
A list of character vectors containing the color palettes. Discrete
palette vectors contain names for each color.
}
\usage{
wacolors
}
\description{
A collection of colorblind-friendly color palettes for various settings in
the state of Washington.
}
\details{
Available palettes (use \code{\link[=wa_pal]{wa_pal()}} to visualize):
\itemize{
\item \code{rainier}
\item \code{forest}
\item \code{washington_pass}
\item \code{palouse}
\item \code{uw}
\item \code{larch}
\item \code{fort_worden}
\item \code{coast}
\item \code{flag}
\item \code{sound_sunset}
\item \code{ferries}
\item \code{forest_fire}
\item \code{sea}
\item \code{sea_star}
\item \code{volcano}
\item \code{baker}
\item \code{mountains}
\item \code{foothills_winter}
\item \code{mountains_sound}
\item \code{vantage}
}
}
\examples{
wacolors$rainier
wacolors$palouse[1:4]

}
\keyword{datasets}
