% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wacolors}
\alias{wacolors}
\title{Washington State Color Palettes}
\format{
A list of objects of type \code{palette}, which are themselves lists of
colors.
}
\usage{
wacolors
}
\description{
A collection of colorblind-friendly color palettes for various settings in
the state of Washington.
}
\keyword{datasets}
